Tensor shape format is done, moving on to the model.
NOTE that this code will have to be re-organized a little bit. Moreover,
we may not necessary have an analogue to "videos." Videos in this case refer to
the task that people are doing. We train seperately on seperate tasks. We may
or may not want to batch. Regardless, we may jsut focus on a specific task for
the demo (i.e. coding vs. email). We need to set up an EXPERIMENTAL PROCEDURE
to decrease the variation in our setup.